backend/requirements.txt
Django==4.2
djangorestframework
pandas
django-cors-headers
reportlab
